+++ packages/node/tests/hashgraphVizualizer.test.ts
@@ -1,9 +1,8 @@
+import { SemanticsType, ActionType, HashGraph, DrpType, newVertex } from "@ts-drp/object";
 import { type Vertex } from "@ts-drp/types";
 import { describe, test, expect, vi, beforeEach } from "vitest";
 
-import { SemanticsType, ActionType, HashGraph, DrpType } from "../src/index.js";
-import { newVertex } from "../src/index.js";
-import { HashGraphVizualizer } from "../src/utils/hashgraphVizualizer.js";
+import { HashGraphVizualizer } from "../src/cli/hashgraphVizualizer.js";
 
 describe("hashGraphVizualizer tests", () => {
 	let hashgraph: HashGraph;
